+++ benchmarks/profiler/deploy/profile_sla_moe_job.yaml
@@ -57,5 +57,5 @@ spec:
       volumes:
         - name: output-volume
           persistentVolumeClaim:
-            claimName: dynamo-pvc
+            claimName: planner-pvc
   backoffLimit: 0
