+++ components/backends/sglang/deploy/disagg_planner.yaml
@@ -22,28 +22,9 @@ spec:
       envFromSecret: hf-token-secret
       componentType: planner
       replicas: 1
-      livenessProbe:
-        exec:
-          command:
-            - /bin/sh
-            - -c
-            - "exit 0"
-        periodSeconds: 60
-        timeoutSeconds: 30
-        failureThreshold: 10
-      readinessProbe:
-        exec:
-          command:
-            - /bin/sh
-            - -c
-            - "exit 0"
-        initialDelaySeconds: 60
-        periodSeconds: 60
-        timeoutSeconds: 30
-        failureThreshold: 10
       pvc:
         create: false
-        name: dynamo-pvc # Must be pre-created before deployment and SLA profiler must have been run
+        name: planner-pvc # Must be pre-created before deployment and SLA profiler must have been run
         mountPoint: /data
       extraPodSpec:
         mainContainer:
@@ -54,7 +35,7 @@ spec:
             - -c
           args:
             - >-
-              python3 -m planner_sla
+              exec python3 -m planner_sla
               --environment=kubernetes
               --backend=sglang
               --adjustment-interval=60
