+++ deploy/utils/manifests/pvc.yaml
@@ -3,11 +3,11 @@
 apiVersion: v1
 kind: PersistentVolumeClaim
 metadata:
-  name: dynamo-pvc
+  name: planner-pvc
   namespace: ${NAMESPACE}
 spec:
   accessModes:
-    - ReadWriteOnce
+    - ReadWriteMany
   resources:
     requests:
       storage: 50Gi
